Dependencies: Add support for Python 3.10 and 3.11
In Python 3.11, the `read_text` and `path` methods of the
`importlib.resources` package of the standard library are deprecated.
They should now be accessed through the `files` interface. This has only
been added since Python 3.9 though, so for older versions we require the
`importlib_resources` backport package to be installed.